+++ api/internal/models/dtos/userDto.go
@@ -3,5 +3,5 @@ package dtos
 type UserDto struct {
 	Id        string  `json:"id"`
 	Email     string  `json:"email"`
-	AnimeIds  []int64 `json:"animeIds"`
+	AnimeIds  []uint `json:"animeIds"`
 }
